tools/pyproject.toml
[tool.poetry]
name = "ipfs-link-tester"
version = "0.1.0"
description = ""
authors = ["a-ok123 <alexey@pastel.network>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
psycopg2-binary = "^2.9.5"
pydantic = "^1.4"
python-dotenv = "^1.0.0"
boto3 = "^1.28.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
